+++ app/src/main/java/com/jk/gogit/network/api/ILogin.kt
@@ -1,7 +1,7 @@
 package com.jk.gogit.network.api
 
-import com.jk.gogit.model.login.AccessToken
-import com.jk.gogit.model.login.AuthRequestModel
+import com.jk.gogit.ui.login.data.response.AccessToken
+import com.jk.gogit.ui.login.data.request.AuthRequestModel
 import retrofit2.http.Body
 import retrofit2.http.Headers
 import retrofit2.http.POST
