+++ app/src/main/java/com/jk/gogit/ui/login/data/response/AccessToken.kt
@@ -1,4 +1,4 @@
-package com.jk.gogit.model.login
+package com.jk.gogit.ui.login.data.response
 
 import com.google.gson.annotations.SerializedName
